+++ changelog.md
@@ -9,6 +9,7 @@
 
 - Updates since 1.2:
     - CanJS 2.0
+    - CanJS 2.0 + RequireJS
 
 ## 1.2 - 2013-08-06